\name{PMV-package}
\alias{PMV-package}
\alias{PMV}
\docType{package}
\title{
What the package does (short line)
}
\description{
More about what it does (maybe more than one line)
}
\details{
\tabular{ll}{
Package: \tab PMV\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-09-23\cr
License: \tab GPL (>= 2)\cr
}
}
\author{
Who wrote it

Maintainer: Who to complain to <yourfault@somewhere.net>
}
\examples{
## test
}
